requirements.txt
# Web框架
Flask==3.1.0
Flask-SQLAlchemy==3.1.1

# 数据库
PyMySQL==1.1.1

# AI和机器学习
openai==1.66.3
ollama==0.4.8

# 向量数据库
chromadb==0.6.3

# 搜索引擎
elasticsearch7==7.17.12

# PDF处理
pdfminer.six==20240706

# 中文处理
jieba==0.42.1

# 数据处理
numpy==2.2.4

# HTTP请求
requests==2.32.3

# 其他工具
python-dotenv==1.0.1
sentence-transformers==5.0.0
opencc==1.1.9
dotenv==0.9.9
